service/llm-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# RAG 및 벡터 검색
faiss-cpu==1.7.4
numpy==1.24.3

# OpenAI API
openai==1.12.0

# HTTP 클라이언트
requests==2.31.0

# 데이터 검증 및 설정
pydantic==2.7.1
python-dotenv==1.0.1

# 유틸리티
python-multipart==0.0.6
